+++ src/apps/Order/components/Main.tsx
@@ -2,51 +2,9 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { SearchLabel } from '@/components/common';
 import { CheckSide } from './CheckSide';
+import { Constants } from '../constants';
 
-type tagsType = '套餐' | '肉品' | '沙拉' | '飲料' | '海鮮' | '人氣單點';
-
-type ProductItemType = {
-  id: number;
-  name: string;
-  type: '單一' | '套餐';
-  tags: tagsType | tagsType[];
-  price: number;
-  image: string;
-  sortNo: number;
-};
-
-export const Main = () => {
-  const productList: ProductItemType[] = [
-    {
-      id: 1,
-      name: 'A5 日本和牛套餐',
-      type: '單一',
-      tags: '肉品',
-      price: 2200,
-      image: 'https://picsum.photos/220/180',
-      sortNo: 1,
-    },
-    {
-      id: 3,
-      name: '豪華全牛套餐',
-      type: '單一',
-      tags: '肉品',
-      price: 1980,
-      image: 'https://picsum.photos/250/280',
-      sortNo: 4,
-    },
-    {
-      id: 8,
-      name: '經典霜降牛套餐',
-      type: '單一',
-      tags: '肉品',
-      price: 1880,
-      image: 'https://picsum.photos/320/180',
-      sortNo: 9,
-    },
-  ];
-
-  return (
+export const Main = () => (
     <div className={clsx(
       'flex flex-row',
       'h-full space-x-6 bg-highlight pl-6 pr-8',
@@ -63,7 +21,7 @@ export const Main = () => {
         <SearchLabel placeholder='餐點名稱' />
         <ul className='mt-6 flex flex-row flex-wrap items-start justify-between gap-6'>
           {
-            productList.map((menu, i) => (
+            Constants.MainProductList.map((menu, i) => (
               <li
                 key={i}
                 className={clsx(
@@ -90,7 +48,6 @@ export const Main = () => {
       </main>
       <CheckSide />
     </div>
-  );
-};
+);
 
 export default Main;
